@karpathy 2024-04-12
(sounds really great! processing through this now)