@karpathy 2024-04-12
I think we'll want to break this up into chunks, a lot of really good stuff here.