@tantara 2024-04-13
One more data point for your PR. I ran your PR on my 4090 (torch==2.2.2, cuda 12.1, nvidia driver 530.30.02)

* llm.c (main branch) : step 1: train loss 4.406586 (took 37.426351 ms)
* llm.c (PR#89) : step 1: train loss 4.406351 (took 29.146125 ms)
* pytorch : iteration 1, loss: 4.579084396362305, time: 33.262ms